+++ src/it/resources/ast/union/2.msg
@@ -1,1 +1,1 @@
-2. should have many union operators
+2. should have many union expression.operators
