+++ src/it/resources/expressions_trees_matches/depth_closure/5.tp
@@ -3,5 +3,5 @@
         @B
         D
 )#
-            C
-                K
+C
+    K
