+++ src/it/resources/expressions_trees_matches/union/2.msg
@@ -1,1 +1,1 @@
-2. should have many union operators
+2. should have many union expression.operators
