+++ src/main/kotlin/expression/operators/DepthTermFunction.kt
@@ -1,4 +1,4 @@
-package operators
+package expression.operators
 
 fun depthTermFunction(node: VisitorFunction): VisitorFunction = { targetTreeNode ->
     VisitorFunctionResponseFactory.createResponseWithDepthTerm(node.invoke(targetTreeNode), targetTreeNode?.moveToFirstChild(), targetTreeNode)
