+++ src/main/kotlin/expression/operators/NodeFunction.kt
@@ -1,4 +1,4 @@
-package operators
+package expression.operators
 
 fun nodeFunction(grammarNodeName: String): VisitorFunction = { targetTreeNode ->
     when (targetTreeNode?.name) {
