+++ src/main/kotlin/expression/operators/VisitorFunction.kt
@@ -1,4 +1,4 @@
-package operators
+package expression.operators
 
 import grammars.ast.ASTNode
 import grammars.ast.BreadthClosure
