+++ src/main/kotlin/grammars/antlr/tree_format/TreeFormatVisitorImplementation.kt
@@ -4,8 +4,8 @@ import org.antlr.v4.runtime.tree.ErrorNode
 import org.antlr.v4.runtime.tree.ParseTree
 import org.antlr.v4.runtime.tree.RuleNode
 import org.antlr.v4.runtime.tree.TerminalNode
-import target_tree.ImpTargetTreeNode
 import target_tree.TargetTreeNode
+import target_tree.default_tree.DefaultTargetTreeNode
 import tree_format.TreeFormatParser
 import tree_format.TreeFormatVisitor
 
@@ -16,9 +16,10 @@ class TreeFormatVisitorImplementation : TreeFormatVisitor<TargetTreeNode> {
     override fun visitChild(context: TreeFormatParser.ChildContext): TargetTreeNode = context.sibling().accept(this)
 
     override fun visitNode(context: TreeFormatParser.NodeContext): TargetTreeNode =
-        ImpTargetTreeNode(context.name.text, context.tag.text, idCount++)
+        DefaultTargetTreeNode(context.name.text, context.tag.text, idCount++)
 
-    override fun visitInformation(context: TreeFormatParser.InformationContext): TargetTreeNode = ImpTargetTreeNode()
+    override fun visitInformation(context: TreeFormatParser.InformationContext): TargetTreeNode =
+        DefaultTargetTreeNode()
 
     override fun visit(parseTree: ParseTree): TargetTreeNode = parseTree.accept(this)
 
@@ -32,12 +33,12 @@ class TreeFormatVisitorImplementation : TreeFormatVisitor<TargetTreeNode> {
         throw NotImplementedError("This method is not supported.")
 
     override fun visitSubtree(context: TreeFormatParser.SubtreeContext): TargetTreeNode {
-        val node: ImpTargetTreeNode = context.node().accept(this) as ImpTargetTreeNode
+        val node: DefaultTargetTreeNode = context.node().accept(this) as DefaultTargetTreeNode
         if (context.child() == null) {
             return node
         }
         node.children = context.child().accept(this).children
-        node.children.forEach { (it as ImpTargetTreeNode).parent = node }
+        node.children.forEach { (it as DefaultTargetTreeNode).parent = node }
         return node
     }
 
@@ -45,6 +46,6 @@ class TreeFormatVisitorImplementation : TreeFormatVisitor<TargetTreeNode> {
         val siblings = context.subtree().toList()
         val siblingNodes = mutableListOf<TargetTreeNode>()
         siblings.forEach { siblingNodes.add(it.accept(this)) }
-        return ImpTargetTreeNode(children = siblingNodes)
+        return DefaultTargetTreeNode(children = siblingNodes)
     }
 }
