+++ src/main/kotlin/grammars/antlr/treepat/Treepat.g4
@@ -1,5 +1,5 @@
-// Define a grammars called Treepat.
-grammars Treepat;
+// Define a grammar called Treepat.
+grammar Treepat;
 
 // Here starts the implementation for INDENT and DEDENT tokens.
 tokens { INDENT, DEDENT }
