+++ src/main/kotlin/target_tree/TargetTree.kt
@@ -1,7 +1,7 @@
 package target_tree
 
 import expression.TreepatExpression
-import operators.VisitorFunctionResponse
+import expression.operators.VisitorFunctionResponse
 
 interface TargetTree<T : TargetTreeNode> {
     var root: T?
@@ -9,9 +9,7 @@ interface TargetTree<T : TargetTreeNode> {
     fun findMatchesRaw(treepatExpression: TreepatExpression): VisitorFunctionResponse =
         treepatExpression.executeExpression(root)
 
-
     fun hasMatch(treepatExpression: TreepatExpression): Boolean = findMatchesRaw(treepatExpression).hasMatch
     fun findMatches(treepatExpression: TreepatExpression): List<List<TargetTreeNode>> =
         findMatchesRaw(treepatExpression).responses.map { it.matches }
 }
-
