+++ src/main/kotlin/target_tree/default_tree/DefaultTargetTree.kt
@@ -1,14 +1,17 @@
-package target_tree
+package target_tree.default_tree
 
 import grammars.antlr.tree_format.TreeFormatVisitorImplementation
+import java.nio.file.Path
 import org.antlr.v4.runtime.CharStreams
 import org.antlr.v4.runtime.CommonTokenStream
 import org.antlr.v4.runtime.tree.ParseTree
+import target_tree.TargetTree
+import target_tree.TargetTreeNode
 import tree_format.TreeFormatLexer
 import tree_format.TreeFormatParser
-import java.nio.file.Path
 
-class DefaultTargetTree<T: TargetTreeNode>(override var root: T? = null) : TargetTree<T> {
+class DefaultTargetTree<T : TargetTreeNode>(override var root: T? = null) :
+    TargetTree<T> {
 
     /**
      * Read tree from grammar.antlr.
